# Faster vartime division for Uint