@andrewwhitehead 2024-06-13
It looks like `Uint::mul_mod` would also be about 70% faster using this update and `split_mut`/`rem_wide_vartime`. The current implementation uses `MontyParams::new_vartime` so it seems like it is vartime for the modulus, but maybe that should be renamed `mul_mod_vartime` anyway? Accepting a `&NonZero<Uint>` for the modulus would also simplify things.