@tarcieri 2024-07-26
> The current implementation uses MontyParams::new_vartime so it seems like it is vartime for the modulus, but maybe that should be renamed mul_mod_vartime anyway? 

That's a good point... perhaps there should be both `Uint::mul_mod` as well as `Uint::mul_mod_vartime` with the former using `MontyParams::new`? (looks like it would need some extra bounds)