@tarcieri 2024-07-26
@fjarri thoughts on this versus #511?