@fjarri 2024-07-27
This is definitely much faster than what I have there, and as for the naming stuff, I'll make a separate PR.