@tarcieri 2024-07-27
This splits `Uint::mul_mod` and `Uint::mul_mod_vartime`: https://github.com/RustCrypto/crypto-bigint/pull/623