@fjarri 2024-08-01
Just a note: I experimented a little with trying to piggyback `rem_wide_vartime()` on `div_rem_vartime()`, and the following kinda works:
```rust
    pub const fn rem_wide_vartime<const R: usize>(lower_upper: (Self, Self), rhs: &NonZero<Self>) -> Self
    where
        Self: Concat<Output=Uint<R>>,
        Uint<R>: Split<Output=Self>,
    {
        let (lo, hi) = lower_upper;
        let wide_self: <Self as Concat>::Output = lo.concat(&hi);
        let (_q, r) = wide_self.div_rem_vartime(rhs);
        r
    }
```
The problem is, it then requires an `<R>` generic parameter in some `Monty` trait methods, which doesn't make sense when they're implemented for `BoxedUint`. So yeah, we'll have to wait until const traits are a thing so that we can say "`Concat::Output` is an `Uint` (or `Integer`)" without spelling out the concrete type.